fix: Next Non Suspense Boundary Error
