+++ libs/client/ParamProvider.tsx
@@ -1,0 +1,16 @@
+import { useEffect } from 'react';
+import { useSearchParams } from 'next/navigation';
+
+export function ParamProvider() {
+  const param = useSearchParams();
+
+  useEffect(() => {
+    // invite-code가 있다면 세션에 저장
+    const inviteCode = param.get('referer');
+    if (inviteCode) {
+      sessionStorage.setItem('invite-code', inviteCode);
+    }
+  }, [param]);
+
+  return <></>;
+}
